Ejercicio 3 del tema 3 actualizado

+++ TEMA-03/JabrEjercicio03Tema3.java
@@ -1,4 +1,4 @@
-/*
+/**
  * Realiza un conversor de pesetas a euros. La cantidad de euros que se quiere
  * convertir debe ser introducida por teclado.
  * 
@@ -6,15 +6,13 @@
  * 
  * 
  */
-
-
 public class JabrEjercicio03Tema3 {
   public static void main (String[] args) {
     System.out.print("Introduce el valor de euros que quieres conventir: ");
     double valorConvertir = Double.parseDouble(System.console().readLine() );;
     double valorPeseta = 166.38;
     
-    double resultado = (valorConvertir * valorPeseta);
-    System.out.printf(" %.2f euros en pesetas son: %.2f pesetas" ,valorConvertir ,resultado);
+    int resultado = (int)(valorConvertir * valorPeseta);
+    System.out.printf(" %.2f euros en pesetas son: %d pesetas" ,valorConvertir ,resultado);
 	}
 }
